requirements.txt
numpy==1.23.0
hydra-core
wandb
plotly
seaborn
tensordict
envpool==0.8.2
tqdm
gym==0.23.1
stable_baselines3
wandb
ray
tensorboard
ipykernel
ipython
einops
pytorch_lightning
gymnasium==0.28.1
gymnasium-notices==0.0.1
mujoco==2.3.3
mujoco-py==2.1.2.14
git+https://github.com/Farama-Foundation/d4rl@master#egg=d4rl